To print the binary representation of integer numbers

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -7,7 +7,7 @@
 
 void print_binary(unsigned long int n)
 {
-	unsigned int copy = n;
+	unsigned long int copy = n;
 
 	copy >> 1;
 	if (copy)
